Rename to ForkPool #98
