--- src/Common/ForkControl/Exception/ForkControlException.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Gaming\Common\ForkControl\Exception;
-
-use Exception;
-
-final class ForkControlException extends Exception
-{
-}
