+++ src/Common/ForkPool/Exception/ForkPoolException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Gaming\Common\ForkPool\Exception;
+
+use Exception;
+
+final class ForkPoolException extends Exception
+{
+}
