+++ src/Common/ForkPool/Wait.php
@@ -2,30 +2,30 @@
 
 declare(strict_types=1);
 
-namespace Gaming\Common\ForkControl;
+namespace Gaming\Common\ForkPool;
 
 final class Wait
 {
     public function __construct(
-        private readonly ForkControl $forkControl,
+        private readonly ForkPool $forkPool,
         private readonly Processes $processes
     ) {
     }
 
-    public function all(): ForkControl
+    public function all(): ForkPool
     {
         while (($processId = pcntl_wait($status)) !== -1) {
             $this->processes->remove($processId);
         }
 
-        return $this->forkControl;
+        return $this->forkPool;
     }
 
-    public function any(): ForkControl
+    public function any(): ForkPool
     {
         $this->processes->remove(pcntl_wait($status));
 
-        return $this->forkControl;
+        return $this->forkPool;
     }
 
     public function killAllWhenAnyExits(int $signal): void
